requirements.txt
tensorflow-gpu==1.2.1
numpy==1.13.1
